Create SPI.v
The objected for this lab is to design a SPI slave core and be able to read/write from and to SPI Core show in figure 1. Also, to learn a full duplex SPI core transmits and receives data to the SPI master through simulation. The SPI core is a 24-bit SPI transfer with input sclk, clk, ssn, mosi, and 16-bit rdata. The outputs will be msio, wstrobe, 16 bits wdata and 7 bits address.